+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -554,6 +554,24 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  @SneakyThrows
+  public void shouldNotUpdateInstanceIfNoChanges() {
+    var id = UUID.randomUUID();
+    createInstance(nod(id));
+    var instance = getById(id).getJson();
+    instanceMessageChecks.createdMessagePublished(id.toString());
+
+    assertThat(update(instance).getStatusCode(), is(204));
+
+    var updatedInstance = getById(id).getJson();
+    //assert that there was no update in database
+    assertThat(updatedInstance.getString("_version"), is("1"));
+    var kafkaEvents = KAFKA_CONSUMER.getMessagesForInstance(id.toString());
+    //assert that there's only CREATE kafka message, no updates
+    assertThat(kafkaEvents.size(), is(1));
+  }
+
+  @Test
   public void cannotProvideAdditionalPropertiesInInstance()
     throws InterruptedException, TimeoutException, ExecutionException {