+++ src/test/java/org/folio/rest/support/messages/ItemEventMessageChecks.java
@@ -6,6 +6,9 @@ import static org.folio.rest.support.AwaitConfiguration.awaitAtMost;
 import static org.folio.services.domainevent.CommonDomainEventPublisher.NULL_ID;
 import static org.folio.utility.ModuleUtility.vertxUrl;
 import static org.folio.utility.RestUtility.TENANT_ID;
+import static org.hamcrest.CoreMatchers.hasItem;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.Matchers.hasProperty;
 
 import io.vertx.core.json.JsonObject;
 import java.util.UUID;
@@ -50,6 +53,11 @@ public class ItemEventMessageChecks {
     awaitAtMost().until(() -> kafkaConsumer.getMessagesForItem(itemId),
       EVENT_MESSAGE_MATCHERS.hasCreateEventMessageFor(
         addInstanceIdToItem(item, instanceId)));
+  }
+
+  public void createdMessagePublished(String itemId) {
+    awaitAtMost().until(() -> kafkaConsumer.getMessagesForItem(itemId),
+      hasItem(hasProperty("type", is("CREATE"))));
   }
 
   public void updatedMessagePublished(JsonObject oldItem, JsonObject newItem) {